% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_hide.R
\name{ogc_hide}
\alias{ogc_hide}
\title{Sortable Table}
\usage{
ogc_hide(x, align, show = 1, label = NULL, vline = "", striped = FALSE)
}
\arguments{
\item{show}{the layer to show by default.
The first layer contains all the date, the second layer hide the second column, etc.}
}
\description{
Sortable Table
}
