% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogc_table.R
\name{ogc_table}
\alias{ogc_table}
\title{Create OGC table with `Tiks`}
\usage{
ogc_table(x, align, label = NULL, vline = "", striped = FALSE)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{align, label, vline}{see knitr::kable.}

\item{striped}{boolean}
}
\description{
Create OGC table with `Tiks`
}
